refactor: extract input_src.rs
